fix(frontend): enhance RHF hooks logic
