+++ frontend/src/components/Menu/MenuForm.tsx
@@ -15,7 +15,7 @@ import { Input } from "@/app-components/inputs/Input";
 import { ToggleableInput } from "@/app-components/inputs/ToggleableInput";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -59,7 +59,7 @@ export const MenuForm: FC<ComponentFormProps<MenuFormData>> = ({
     formState: { errors },
     resetField,
     handleSubmit,
-  } = useForm<IMenuItemAttributes>({
+  } = useStrictForm<IMenuItemAttributes>({
     defaultValues: DEFAULT_VALUES,
     rules: {
       type: {
