+++ frontend/src/components/categories/CategoryForm.tsx
@@ -13,7 +13,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType, RouterType } from "@/services/types";
@@ -47,7 +47,7 @@ export const CategoryForm: FC<ComponentFormProps<ICategory>> = ({
     register,
     formState: { errors },
     handleSubmit,
-  } = useForm<ICategoryAttributes>({
+  } = useStrictForm<ICategoryAttributes>({
     rules: {
       label: { required: t("message.label_is_required") },
     },
