+++ frontend/src/components/content-types/ContentTypeForm.tsx
@@ -15,7 +15,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -40,7 +40,7 @@ export const ContentTypeForm: FC<ComponentFormProps<IContentType>> = ({
     setValue,
     formState: { errors },
     handleSubmit,
-  } = useForm<Partial<IContentType>>({
+  } = useStrictForm<Partial<IContentType>>({
     defaultValues: {
       name: contentType?.name || "",
       fields: contentType?.fields || FIELDS_FORM_DEFAULT_VALUES,
