+++ frontend/src/components/contents/ContentForm.tsx
@@ -21,7 +21,7 @@ import { Adornment } from "@/app-components/inputs/Adornment";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -141,7 +141,7 @@ export const ContentForm: FC<ComponentFormProps<IContent, IContentType>> = ({
     control,
     formState: { errors },
     handleSubmit,
-  } = useForm<IContentAttributes & { [key: string]: any }>({
+  } = useStrictForm<IContentAttributes & { [key: string]: any }>({
     defaultValues: {
       entity: content?.entity || "",
       status: content?.status || false,
