+++ frontend/src/components/context-vars/ContextVarForm.tsx
@@ -14,7 +14,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -55,7 +55,7 @@ export const ContextVarForm: FC<ComponentFormProps<IContextVar>> = ({
     setValue,
     formState: { errors },
     handleSubmit,
-  } = useForm<IContextVarAttributes>({
+  } = useStrictForm<IContextVarAttributes>({
     defaultValues: {
       name: contextVar?.name || "",
       label: contextVar?.label || "",
