+++ frontend/src/components/labels/LabelForm.tsx
@@ -12,7 +12,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -46,7 +46,7 @@ export const LabelForm: FC<ComponentFormProps<ILabel>> = ({
     setValue,
     formState: { errors },
     handleSubmit,
-  } = useForm<ILabelAttributes>({
+  } = useStrictForm<ILabelAttributes>({
     defaultValues: {
       name: label?.name || "",
       title: label?.title || "",
