+++ frontend/src/components/languages/LanguageForm.tsx
@@ -14,7 +14,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -47,7 +47,7 @@ export const LanguageForm: FC<ComponentFormProps<ILanguage>> = ({
     formState: { errors },
     handleSubmit,
     control,
-  } = useForm<ILanguageAttributes>({
+  } = useStrictForm<ILanguageAttributes>({
     defaultValues: {
       title: language?.title || "",
       code: language?.code || "",
