+++ frontend/src/components/nlp/components/NlpEntityForm.tsx
@@ -19,7 +19,7 @@ import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import { Input } from "@/app-components/inputs/Input";
 import { useCreate } from "@/hooks/crud/useCreate";
 import { useUpdate } from "@/hooks/crud/useUpdate";
-import { useForm } from "@/hooks/useForm";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType } from "@/services/types";
@@ -55,7 +55,7 @@ export const NlpEntityVarForm: FC<ComponentFormProps<INlpEntity>> = ({
     register,
     formState: { errors },
     handleSubmit,
-  } = useForm<INlpEntityAttributes>({
+  } = useStrictForm<INlpEntityAttributes>({
     defaultValues: {
       name: nlpEntity?.name || "",
       doc: nlpEntity?.doc || "",
