+++ frontend/src/components/users/EditUserForm.tsx
@@ -8,12 +8,13 @@
 
 import { Button, Grid, Link } from "@mui/material";
 import { FC, Fragment, useEffect } from "react";
-import { Controller, useForm } from "react-hook-form";
+import { Controller } from "react-hook-form";
 
 import { ContentContainer, ContentItem } from "@/app-components/dialogs";
 import AutoCompleteEntitySelect from "@/app-components/inputs/AutoCompleteEntitySelect";
 import { Input } from "@/app-components/inputs/Input";
 import { useUpdate } from "@/hooks/crud/useUpdate";
+import { useStrictForm } from "@/hooks/useStrictForm";
 import { useToast } from "@/hooks/useToast";
 import { useTranslate } from "@/hooks/useTranslate";
 import { EntityType, Format } from "@/services/types";
@@ -46,7 +47,7 @@ export const EditUserForm: FC<ComponentFormProps<IUser, IRole[]>> = ({
     control,
     formState: { errors },
     handleSubmit,
-  } = useForm<IUserAttributes>({
+  } = useStrictForm<IUserAttributes>({
     defaultValues: { roles: roles?.map((role) => role.id) },
   });
   const validationRules = {
